refactor(web): 공통 컴포넌트 이전
